+++ lib/ansible/cli/galaxy.py
@@ -16,8 +16,8 @@ from jinja2 import Environment, FileSystemLoader
 
 import ansible.constants as C
 from ansible import context
-from ansible.arguments import optparse_helpers as opt_help
 from ansible.cli import CLI
+from ansible.cli.arguments import optparse_helpers as opt_help
 from ansible.errors import AnsibleError, AnsibleOptionsError
 from ansible.galaxy import Galaxy
 from ansible.galaxy.api import GalaxyAPI
