+++ lib/ansible/context.py
@@ -15,14 +15,18 @@ running the ansible command line tools.
 These APIs are still in flux so do not use them unless you are willing to update them with every Ansible release
 """
 
-from ansible.arguments.context_objects import CLIArgs, GlobalCLIArgs
+from ansible.utils.context_objects import CLIArgs, GlobalCLIArgs
 
+
+__all__ = ('CLIARGS',)
 
 # Note: this is not the singleton version.  The Singleton is only created once the program has
 # actually parsed the args
 CLIARGS = CLIArgs({})
 
 
+# This should be called immediately after cli_args are processed (parsed, validated, and any
+# normalization performed on them).  No other code should call it
 def _init_global_context(cli_args):
     """Initialize the global context objects"""
     global CLIARGS
