+++ lib/ansible/utils/context_objects.py
@@ -1,4 +1,4 @@
-# Copyright: (c) 2018, Toshio Kuratomi <tkuratomi@ansible.com>
+# Copyright: (c) 2018, Ansible Project
 # GNU General Public License v3.0+ (see COPYING or https://www.gnu.org/licenses/gpl-3.0.txt)
 
 # Make coding more python3-ish
