+++ test/units/executor/test_task_queue_manager_callbacks.py
@@ -22,10 +22,10 @@ from units.compat import unittest
 from units.compat.mock import MagicMock
 
 from ansible import context
-from ansible.arguments import context_objects as co
 from ansible.executor.task_queue_manager import TaskQueueManager
 from ansible.playbook import Playbook
 from ansible.plugins.callback import CallbackBase
+from ansible.utils import context_objects as co
 
 __metaclass__ = type
